requirements.txt
yt-dlp>=2024.1.1
google-genai>=0.1.0
requests>=2.31.0
pyyaml>=6.0
pillow>=10.0.0
pytest>=7.4.0
rich>=13.0.0
black>=23.0.0
flake8>=6.0.0